Enterprise text_rag_raw/requirements.txt
openai
python-dotenv
tiktoken
numpy
faiss-cpu
tqdm
pypdf
fastapi
uvicorn[standard]
pydantic